// src/core/types/ProtocolParams.ts
export interface PlutusV2ScriptParams extends PlutusV1ScriptParams {
  readonly 'serialiseData-cpu-arguments-intercept': number;
  readonly 'serialiseData-memory-arguments-intercept': number;
  readonly 'verifySchnorrSecp256k1Signature-cpu-arguments-slope': number;
  readonly 'verifyEcdsaSecp256k1Signature-memory-arguments': number;
  readonly 'verifyEcdsaSecp256k1Signature-cpu-arguments': number;
  readonly 'verifySchnorrSecp256k1Signature-cpu-arguments-intercept': number;
  readonly 'verifySchnorrSecp256k1Signature-memory-arguments': number;
  readonly 'serialiseData-memory-arguments-slope': number;
  readonly 'serialiseData-cpu-arguments-slope': number;
}
export interface ProtocolParams {
  readonly network: 'mainnet' | 'preview' | 'preprod';
  readonly protocolVersion: {
    readonly major: number;
    readonly minor: number;
  };
  readonly maxTxSize: bigint;
  readonly txFeeFixed: bigint;
  readonly txFeePerByte: bigint;
  readonly minUTxOValue: bigint;
  readonly coinsPerUtxoByte: bigint;
  readonly costModels: {
    readonly PlutusScriptV1: PlutusV1ScriptParams;
    readonly PlutusScriptV2: PlutusV2ScriptParams;
  };
  readonly executionUnitPrices: {
    readonly priceSteps: number;
    readonly priceMemory: number;
  };
  readonly maxTxExecutionUnits: {
    readonly steps: bigint;
    readonly memory: bigint;
  };
  readonly maxValueSize: bigint;
  readonly collateralPercentage: number;
  readonly maxCollateralInputs: number;
}